+++ src/MvvmBlazor/Components/MvvmOwningComponentBase.cs
@@ -1,9 +1,4 @@
 ﻿using Microsoft.AspNetCore.Components;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MvvmBlazor.Components
 {
