+++ src/MvvmBlazor/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using MvvmBlazor.Components;
 using MvvmBlazor.Internal.Bindings;
 using MvvmBlazor.Internal.Parameters;
 using MvvmBlazor.Internal.WeakEventListener;
@@ -9,11 +10,12 @@ namespace MvvmBlazor.Extensions
     {
         public static IServiceCollection AddMvvm(this IServiceCollection serviceCollection)
         {
-            serviceCollection.AddSingleton<IWeakEventManagerFactory, WeakEventManagerFactory>();
             serviceCollection.AddSingleton<IBindingFactory, BindingFactory>();
             serviceCollection.AddSingleton<IParameterResolver, ParameterResolver>();
             serviceCollection.AddSingleton<IParameterCache, ParameterCache>();
             serviceCollection.AddSingleton<IViewModelParameterSetter, ViewModelParameterSetter>();
+            serviceCollection.AddTransient<IWeakEventManager, WeakEventManager>();
+            serviceCollection.AddTransient<IBinder, Binder>();
             serviceCollection.AddHttpContextAccessor();
 
             return serviceCollection;
